解决内部泛型类在旧版Eclipse或MyEclipse下的编译问题
